requirements.txt
langchain
langchain-core
flask 
sentence-transformers
pypdf
pyMuPDF
python-dotenv
langchain-pinecone 
langchain-openai
langchain-openrouter
langchain-google-genai
langchain-community
-e .